refactor(flui-painting): Mythos Step 5 -- split display_list.rs god module
Split the 2,434-LOC `crates/flui-painting/src/display_list.rs` god
module into a concern-based `display_list/` directory. Six files,
each focused on one concern; total LOC drops to ~2,016 (net -418
from removing duplicate docstring noise and intermediate boilerplate).

Files created:

| File                                       | LOC | Concern                                                                                |
|--------------------------------------------|-----|----------------------------------------------------------------------------------------|
| `display_list/mod.rs`                      | 410 | DisplayList struct + Default + iter/iter_mut + apply_transform/filter/map/to_opacity/clear + push/append + AsRef/AsMut + IntoIterator + Index/IndexMut |
| `display_list/command.rs`                  | 433 | 29-variant `DrawCommand` enum + `CommandKind` enum                                     |
| `display_list/command_ops.rs`              | 745 | `DrawCommand` impl: with_opacity (~250 LOC), bounds (~250 LOC), transform/transform_mut/paint/has_paint/kind/is_*/apply_transform |
| `display_list/sealed.rs`                   | 225 | private::Sealed + DisplayListCore trait + DisplayListExt trait + 4 blanket impls       |
| `display_list/stats.rs`                    | 102 | `DisplayListStats` struct + zero()/new() + Display impl                                |
| `display_list/hit_region.rs`               | 101 | `PointerEvent` + `PointerEventKind` + `HitRegion` + `HitRegionHandler`                 |

The split honors concerns by responsibility:

- mod.rs owns the `DisplayList` struct lifecycle (push, append,
  clear, recalculate_bounds) and the public iteration/index API.
  `commands_mut` stays `pub` here -- U10 demotes it to `pub(crate)`.
- command.rs owns the closed-enum vocabulary (29 variants +
  CommandKind). This is the GPU lowering contract with
  `flui-engine`'s wgpu backend (exhaustive match).
- command_ops.rs owns the per-variant pattern-match operations.
  The largest file in the split; future Outstanding refactor is a
  `macro_rules! gen_command_accessors!`-style collapse mirroring
  the flui-layer chain's Step 4 dispatch macro.
- sealed.rs owns the sealed-extension-trait pair + 4 blanket impls
  (DisplayList, Arc<DisplayList>, Box<DisplayList>, &DisplayList).
  The Arc<DisplayList> blanket impl is load-bearing for
  flui-layer::Layer::Picture retained-layer caching.
- stats.rs owns the structured-count value type.
- hit_region.rs owns the event-handler registry types (pointer
  events recorded here; resolution lives in flui-interaction).

The 3 inline tests from the old display_list.rs are preserved in
`display_list/mod.rs` for this step. U8 of the chain extracts them
to `crates/flui-painting/tests/display_list_command.rs` etc.

Public API unchanged. External callers (`flui-engine::wgpu::backend`,
`wgpu::layer_render`, `wgpu::debug`, `flui-layer::layer::canvas`,
`flui-layer::layer::picture`) compile without changes. The
`pub mod display_list;` declaration in `lib.rs` resolves to the new
directory automatically; the re-exports from `flui_types::painting`
(BlendMode, Paint, etc.) are preserved in `display_list/mod.rs`.

Verification:
- `cargo check -p flui-painting` clean.
- `cargo test -p flui-painting --lib` 64 passing.
- `cargo test -p flui-painting --tests` 9 passing.
- `cargo check --workspace` clean.
- `bash scripts/port-check.sh` 6 triggers ok.

Net unsafe delta: 0 (crate is #[forbid(unsafe_code)]; no unsafe
introduced; none removed).
Net LOC delta in production code: -418 LOC across the split files
vs. the original god module.

Mythos chain step 5 of 14. Next: U6 splits text_layout.rs (1243 LOC)
into a `text_layout/` directory + flattens the unnecessary
`#[cfg(feature = "text")] mod inner` indirection.
